move transfer fixtures to tests_transfers/transfer_fixtures
