+++ tests/test_unit/test_runs/run_fixtures/__init__.py
@@ -1,2 +1,2 @@
-from tests.test_unit.test_runs.run_fixtures.group_run_fixture import group_run
-from tests.test_unit.test_runs.run_fixtures.group_runs_fixture import group_runs
+from tests.test_unit.test_runs.run_fixtures.run_fixture import group_run
+from tests.test_unit.test_runs.run_fixtures.runs_fixture import group_runs
